refactor types and enums
