+++ src/modules/stage2-processing/CacheEnumsAndTypes.ts
@@ -1,0 +1,19 @@
+export enum LookupEnum {
+    POST,
+    COMMUNITY_SEEN_COUNT,
+}
+
+export type Stage2CacheData = {
+    sessId: string;
+    cachedPosts: {
+        string?: boolean;
+    };
+    perCommunitySeenPost: {
+        string?: number;
+    };
+};
+
+export type LookupData = {
+    type: LookupEnum;
+    value: string;
+};
